// src/models/StatsEnum.ts
enum StatsEnum {
  attackDamage = 'attackDamage',
  attackDamageLvl1 = 'attackDamageLvl1',
  attackDamageLvl18 = 'attackDamageLvl18',

  attackSpeed = 'attackSpeed',
  attackSpeedBonus = 'attackSpeedBonus',
  attackSpeedRatio = 'attackSpeedRatio',

  critChance = 'critChance',
  critDamage = 'critDamage',

  lethality = 'lethality',
  armorPenetration = 'armorPenetration',

  magicFlatPenetration = 'magicFlatPenetration',
  magicPenetration = 'magicPenetration',

  health = 'health',
  healthLvl1 = 'healthLvl1',
  healthLvl18 = 'healthLvl18',

  armor = 'armor',
  armorLvl1 = 'armorLvl1',
  armorLvl18 = 'armorLvl18',

  magicResistance = 'magicResistance',
  magicResistanceLvl1 = 'magicResistanceLvl1',
  magicResistanceLvl18 = 'magicResistanceLvl18',
}
export default StatsEnum;